third_party/WebKit/LayoutTests/fast/dom/DOMException/dispatch-event-exception-expected.txt

Tests the properties of the exception thrown by dispatchEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.toString() is "InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable."
PASS Object.prototype.toString.call(e) is "[object DOMException]"
FAIL Object.prototype.toString.call(e.__proto__) should be [object DOMExceptionPrototype]. Was [object Object].
PASS e.constructor.toString() is "function DOMException() { [native code] }"
PASS e.constructor is window.DOMException
PASS successfullyParsed is true

TEST COMPLETE
